Makefile Modified, Code changes as per experiment purpose

+++ code/02_simple_sv2c_return/c/function3.c
@@ -2,5 +2,6 @@
 
 double myCFunction3( int A, float B, double C )
 {
-    return (double)A+(double)(B)+C;
+    // return (double)A+(double)(B)+C;
+    return A+B+C;
 }
